A web-app to track the Student's Attendence in class
A Web-App to track the student's attendence in class, that can save everyday's class attendence to keep a track, how many students were attending the class. This can also be used for tracking passengers, or no. of people running in a race, also to keep a track of guests in a party.